+++ R/package-docs.R
@@ -1,7 +1,4 @@
 
-# package documentation --------------------------------------------------------
-
-
 #' Quick introduction to huxtable
 #'
 #' Huxtable is a package for creating HTML and LaTeX tables. It provides similar